# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Wisdoverse Forge CLI API
  description: |
    REST API for the Wisdoverse Forge platform, covering agent lifecycle management,
    real-time event streaming (SSE), server-side wait, and multi-agent group
    orchestration. This spec documents every endpoint surfaced by the CLI tool (`af`).
  version: 1.0.0
  contact:
    name: Wisdoverse Forge
    url: https://wisdoverse.com
  license:
    name: LicenseRef-Wisdoverse-Forge-BSL-1.1
    url: https://github.com/Wisdoverse/Wisdoverse-Forge/blob/main/LICENSE
servers:
  - url: '{scheme}://{host}:{port}'
    description: Wisdoverse Forge server
    variables:
      scheme:
        default: https
        enum: [http, https]
      host:
        default: localhost
      port:
        default: '4003'
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/agents:
    get:
      operationId: listAgents
      summary: List agents
      description: |
        List agents scoped to the authenticated user's organization.
        Default scope is `visible` (own + project + collaborator agents).
        `scope=all` requires admin role.
      tags: [Agents]
      parameters:
        - name: scope
          in: query
          schema:
            type: string
            enum: [visible, drafts, project, all]
            default: visible
        - name: projectId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Agent list
          content:
            application/json:
              schema:
                type: object
                required: [ok, agents]
                properties:
                  ok:
                    type: boolean
                    const: true
                  agents:
                    type: array
                    items:
                      $ref: '#/components/schemas/AgentData'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      operationId: createAgent
      summary: Create a new agent
      tags: [Agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: Agent created
          content:
            application/json:
              schema:
                type: object
                required: [ok, agent]
                properties:
                  ok:
                    type: boolean
                    const: true
                  agent:
                    $ref: '#/components/schemas/AgentData'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/agents/{id}/restart:
    parameters:
      - $ref: '#/components/parameters/AgentId'
    post:
      operationId: restartAgent
      summary: Restart agent container
      tags: [Agent Actions]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestartAgentRequest'
      responses:
        '200':
          description: Agent restarted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for CLI and programmatic access

  # -------------------------------------------------------------------------
  # Parameters
  # -------------------------------------------------------------------------
  parameters:
    AgentId:
      name: id
      in: path
      required: true
      description: Agent UUID
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  details:
                    type: array
                    items:
                      type: object
                      description: Zod validation issue

  # -------------------------------------------------------------------------
  # Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          const: true
    ErrorResponse:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          const: false
        error:
          type: string
          description: Human-readable error message

    # -- Auth --------------------------------------------------------------
...
    AgentStatus:
      type: string
      enum: [idle, working, waiting, attention, offline]
    GroupRole:
      type: string
      enum: [manager, worker, none]
    CliTool:
      type: string
      enum: [claude, opencode, codex, gemini]
    TokenCount:
      type: object
      required: [current, cumulative]
      properties:
        current:
          type: integer
        cumulative:
          type: integer
    ZonePosition:
      type: object
      required: [q, r]
      properties:
        q:
          type: integer
        r:
          type: integer
    GitStatusSummary:
      type: object
      properties:
        branch:
          type: string
        ahead:
          type: integer
        behind:
          type: integer
        staged:
          type: object
          properties:
            added:
              type: integer
            modified:
              type: integer
            deleted:
              type: integer
        unstaged:
          type: object
          properties:
            added:
              type: integer
            modified:
              type: integer
            deleted:
              type: integer
        untracked:
          type: integer
        totalFiles:
          type: integer
        linesAdded:
          type: integer
        linesRemoved:
          type: integer
        lastCommitTime:
          type: [integer, 'null']
        lastCommitMessage:
          type: [string, 'null']
        isRepo:
          type: boolean
        lastChecked:
          type: integer
    AgentData:
      type: object
      required: [id, name, runtimeId, status, createdAt, lastActivity]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        runtimeId:
          type: string
          description: Runtime identifier (e.g. 'af-a1b2c3d4')
        containerId:
          type: string
          description: Docker container hash
        status:
          $ref: '#/components/schemas/AgentStatus'
        cliSessionId:
          type: string
        createdAt:
          type: integer
          description: Unix timestamp in milliseconds
        lastActivity:
          type: integer
          description: Unix timestamp in milliseconds
        cwd:
          type: string
          description: Working directory (display path, ~/...)
        currentTool:
          type: string
        tokens:
          $ref: '#/components/schemas/TokenCount'
        gitStatus:
          $ref: '#/components/schemas/GitStatusSummary'
        zonePosition:
          $ref: '#/components/schemas/ZonePosition'
        groupId:
          type: string
          format: uuid
        groupRole:
          $ref: '#/components/schemas/GroupRole'
        userId:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        cliTool:
          $ref: '#/components/schemas/CliTool'
        ownerUsername:
          type: string
        ownerEmail:
          type: string
    CreateAgentRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        cwd:
          type: string
        chromeEnabled:
          type: boolean
        skipPermissions:
          type: boolean
        groupId:
          type: string
          format: uuid
        groupRole:
          $ref: '#/components/schemas/GroupRole'
        projectId:
          type: string
          format: uuid
        cliTool:
          $ref: '#/components/schemas/CliTool'
        profileId:
          type: string
          format: uuid
          description: Resource profile ID. If omitted, the org default is used.
...
    RestartAgentRequest:
      type: object
      properties:
        chromeEnabled:
          type: boolean
        skipPermissions:
          type: boolean
...